Jsonify.ts: Renamed IsJsonable to Jsonify. Fixed hard-wrap, use tab-indentation, and simplify description.
